Nesting endorsement classes.
