+++ app/models/journeyDomain/incident/endorsement/EndorsementDomain.scala
@@ -20,7 +20,7 @@ import cats.implicits._
 import models.Index
 import models.journeyDomain.{GettableAsReaderOps, UserAnswersReader}
 import models.reference.Country
-import pages.incident.{EndorsementAuthorityPage, EndorsementCountryPage, EndorsementDatePage, EndorsementLocationPage}
+import pages.incident.endorsement.{EndorsementAuthorityPage, EndorsementCountryPage, EndorsementDatePage, EndorsementLocationPage}
 
 import java.time.LocalDate
