--- app/views/incident/EndorsementLocationView.scala.html
@@ -1,54 +1,0 @@
-@*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import controllers.incident.routes
-@import models.MovementReferenceNumber
-@import viewModels.components.InputTextViewModel._
-@import views.html.components.{Button, ErrorSummary, InputText}
-
-@this(
-    mainTemplate: MainTemplate,
-    formHelper: FormWithCSRF,
-    inputText: InputText,
-    errorSummary: ErrorSummary,
-    button: Button
-)
-
-@(form: Form[String], mrn: MovementReferenceNumber, countryName: String, mode: Mode, index: Index)(implicit request: Request[_], messages: Messages)
-
-@mainTemplate(
-    title = messages("incident.endorsementLocation.title", countryName),
-    showBackLink = true,
-    mrn = Some(mrn)
-) {
-
-    @formHelper(action = routes.EndorsementLocationController.onSubmit(mrn, mode, index), Symbol("autoComplete") -> "off") {
-
-        @errorSummary(form.errors)
-
-        @inputText(
-            field = form("value"),
-            inputTextType = OrdinaryTextInput(
-                heading = messages("incident.endorsementLocation.heading", countryName),
-                caption = Some(messages("incident.caption"))
-            ),
-            hint = Some(messages("incident.endorsementLocation.hintText")),
-            inputClass = InputSize.Width20
-        )
-
-        @button(attributes = Map("id" -> "submit"))
-    }
-}
