+++ app/views/incident/endorsement/EndorsementAuthorityView.scala.html
@@ -14,37 +14,38 @@
  * limitations under the License.
  *@
 
-@import controllers.incident.routes
+@import controllers.incident.endorsement.routes
 @import models.MovementReferenceNumber
-@import views.html.components.{Button, DateErrorSummary, InputDate}
-
-@import java.time.LocalDate
+@import viewModels.components.InputTextViewModel._
+@import views.html.components.{Button, ErrorSummary, InputText}
 
 @this(
     mainTemplate: MainTemplate,
     formHelper: FormWithCSRF,
-    inputDate: InputDate,
-    errorSummary: DateErrorSummary,
+    inputText: InputText,
+    errorSummary: ErrorSummary,
     button: Button
 )
 
-@(form: Form[LocalDate], mrn: MovementReferenceNumber, index: Index, mode: Mode)(implicit request: Request[_], messages: Messages)
+@(form: Form[String], mrn: MovementReferenceNumber, mode: Mode, index: Index)(implicit request: Request[_], messages: Messages)
 
 @mainTemplate(
-    title = messages("incident.endorsementDate.title"),
+    title = messages("incident.endorsement.authority.title"),
     showBackLink = true,
     mrn = Some(mrn)
 ) {
 
-    @formHelper(action = routes.EndorsementDateController.onSubmit(mrn, index, mode), Symbol("autoComplete") -> "off") {
+    @formHelper(action = routes.EndorsementAuthorityController.onSubmit(mrn, mode, index), Symbol("autoComplete") -> "off") {
 
-        @errorSummary(form)
+        @errorSummary(form.errors)
 
-        @inputDate(
+        @inputText(
             field = form("value"),
-            legend = messages("incident.endorsementDate.heading"),
-            hint = Some(messages("incident.endorsementDate.hint")),
-            caption = Some(messages("incident.caption"))
+            inputTextType = OrdinaryTextInput(
+                heading = messages("incident.endorsement.authority.heading"),
+                caption = Some(messages("incident.caption"))
+            ),
+            inputClass = InputSize.Width20
         )
 
         @button(attributes = Map("id" -> "submit"))
