+++ app/views/incident/endorsement/EndorsementDateView.scala.html
@@ -1,0 +1,52 @@
+@*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import controllers.incident.endorsement.routes
+@import models.MovementReferenceNumber
+@import views.html.components.{Button, DateErrorSummary, InputDate}
+
+@import java.time.LocalDate
+
+@this(
+    mainTemplate: MainTemplate,
+    formHelper: FormWithCSRF,
+    inputDate: InputDate,
+    errorSummary: DateErrorSummary,
+    button: Button
+)
+
+@(form: Form[LocalDate], mrn: MovementReferenceNumber, index: Index, mode: Mode)(implicit request: Request[_], messages: Messages)
+
+@mainTemplate(
+    title = messages("incident.endorsement.date.title"),
+    showBackLink = true,
+    mrn = Some(mrn)
+) {
+
+    @formHelper(action = routes.EndorsementDateController.onSubmit(mrn, index, mode), Symbol("autoComplete") -> "off") {
+
+        @errorSummary(form)
+
+        @inputDate(
+            field = form("value"),
+            legend = messages("incident.endorsement.date.heading"),
+            hint = Some(messages("incident.endorsement.date.hint")),
+            caption = Some(messages("incident.caption"))
+        )
+
+        @button(attributes = Map("id" -> "submit"))
+    }
+}
