+++ conf/app.incident.endorsement.routes
@@ -1,0 +1,20 @@
+
+GET        /:mrn/incidents/:index/endorsement-date                    controllers.incident.endorsement.EndorsementDateController.onPageLoad(mrn: MovementReferenceNumber, index: Index, mode: Mode = NormalMode)
+POST       /:mrn/incidents/:index/endorsement-date                    controllers.incident.endorsement.EndorsementDateController.onSubmit(mrn: MovementReferenceNumber, index: Index, mode: Mode = NormalMode)
+GET        /:mrn/incidents/:index/change-endorsement-date             controllers.incident.endorsement.EndorsementDateController.onPageLoad(mrn: MovementReferenceNumber, index: Index, mode: Mode = CheckMode)
+POST       /:mrn/incidents/:index/change-endorsement-date             controllers.incident.endorsement.EndorsementDateController.onSubmit(mrn: MovementReferenceNumber, index: Index, mode: Mode = CheckMode)
+
+GET        /:mrn/incidents/:index/endorsement-authority               controllers.incident.endorsement.EndorsementAuthorityController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = NormalMode, index: Index)
+POST       /:mrn/incidents/:index/endorsement-authority               controllers.incident.endorsement.EndorsementAuthorityController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = NormalMode, index: Index)
+GET        /:mrn/incidents/:index/change-endorsement-authority        controllers.incident.endorsement.EndorsementAuthorityController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = CheckMode, index: Index)
+POST       /:mrn/incidents/:index/change-endorsement-authority        controllers.incident.endorsement.EndorsementAuthorityController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = CheckMode, index: Index)
+
+GET        /:mrn/incidents/:index/endorsement-location                controllers.incident.endorsement.EndorsementLocationController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = NormalMode, index: Index)
+POST       /:mrn/incidents/:index/endorsement-location                controllers.incident.endorsement.EndorsementLocationController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = NormalMode, index: Index)
+GET        /:mrn/incidents/:index/change-endorsement-location         controllers.incident.endorsement.EndorsementLocationController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = CheckMode, index: Index)
+POST       /:mrn/incidents/:index/change-endorsement-location         controllers.incident.endorsement.EndorsementLocationController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = CheckMode, index: Index)
+
+GET        /:mrn/incidents/:index/endorsement-country                 controllers.incident.endorsement.EndorsementCountryController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = NormalMode, index: Index)
+POST       /:mrn/incidents/:index/endorsement-country                 controllers.incident.endorsement.EndorsementCountryController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = NormalMode, index: Index)
+GET        /:mrn/incidents/:index/change-endorsement-country          controllers.incident.endorsement.EndorsementCountryController.onPageLoad(mrn: MovementReferenceNumber, mode: Mode = CheckMode, index: Index)
+POST       /:mrn/incidents/:index/change-endorsement-country          controllers.incident.endorsement.EndorsementCountryController.onSubmit(mrn: MovementReferenceNumber, mode: Mode = CheckMode, index: Index)
