+++ test/controllers/incident/endorsement/EndorsementLocationControllerSpec.scala
@@ -14,96 +14,91 @@
  * limitations under the License.
  */
 
-package controllers.incident
+package controllers.incident.endorsement
 
 import base.{AppWithDefaultMockFixtures, SpecBase}
-import forms.DateFormProvider
+import forms.incident.EndorsementLocationFormProvider
+import generators.Generators
 import models.NormalMode
+import models.reference.Country
 import navigation.IncidentNavigatorProvider
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
-import pages.incident.EndorsementDatePage
+import org.scalacheck.Arbitrary.arbitrary
+import pages.incident.endorsement
+import pages.incident.endorsement.{EndorsementCountryPage, EndorsementLocationPage}
 import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import views.html.incident.EndorsementDateView
+import views.html.incident.endorsement.EndorsementLocationView
 
-import java.time.{Clock, LocalDate, ZoneOffset}
 import scala.concurrent.Future
 
-class EndorsementDateControllerSpec extends SpecBase with AppWithDefaultMockFixtures {
+class EndorsementLocationControllerSpec extends SpecBase with AppWithDefaultMockFixtures with Generators {
 
-  private val minDate = frontendAppConfig.endorsementDateMin
-  private val zone    = ZoneOffset.UTC
-  private val clock   = Clock.systemDefaultZone.withZone(zone)
-
-  private val formProvider              = new DateFormProvider(clock)
-  private val form                      = formProvider("incident.endorsementDate", minDate)
-  private val mode                      = NormalMode
-  private lazy val endorsementDateRoute = routes.EndorsementDateController.onPageLoad(mrn, index, mode).url
-  private val date                      = LocalDate.now
+  private val country                       = arbitrary[Country].sample.value
+  private val formProvider                  = new EndorsementLocationFormProvider()
+  private val form                          = formProvider("incident.endorsement.location", country.description)
+  private val mode                          = NormalMode
+  private lazy val endorsementLocationRoute = routes.EndorsementLocationController.onPageLoad(mrn, mode, index).url
 
   override def guiceApplicationBuilder(): GuiceApplicationBuilder =
     super
       .guiceApplicationBuilder()
       .overrides(bind(classOf[IncidentNavigatorProvider]).toInstance(fakeIncidentNavigatorProvider))
 
-  "EndorsementDate Controller" - {
+  "EndorsementLocation Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
-      setExistingUserAnswers(emptyUserAnswers)
+      val userAnswers = emptyUserAnswers.setValue(EndorsementCountryPage(index), country)
 
-      val request = FakeRequest(GET, endorsementDateRoute)
+      setExistingUserAnswers(userAnswers)
+
+      val request = FakeRequest(GET, endorsementLocationRoute)
 
       val result = route(app, request).value
 
-      val view = injector.instanceOf[EndorsementDateView]
+      val view = injector.instanceOf[EndorsementLocationView]
 
       status(result) mustEqual OK
 
       contentAsString(result) mustEqual
-        view(form, mrn, index, mode)(request, messages).toString
+        view(form, mrn, country.description, mode, index)(request, messages).toString
     }
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val userAnswers = emptyUserAnswers.setValue(EndorsementDatePage(index), date)
+      val userAnswers = emptyUserAnswers
+        .setValue(endorsement.EndorsementCountryPage(index), country)
+        .setValue(EndorsementLocationPage(index), "test string")
       setExistingUserAnswers(userAnswers)
 
-      val request = FakeRequest(GET, endorsementDateRoute)
+      val request = FakeRequest(GET, endorsementLocationRoute)
 
       val result = route(app, request).value
 
-      val filledForm = form.bind(
-        Map(
-          "value.day"   -> date.getDayOfMonth.toString,
-          "value.month" -> date.getMonthValue.toString,
-          "value.year"  -> date.getYear.toString
-        )
-      )
+      val filledForm = form.bind(Map("value" -> "test string"))
 
-      val view = injector.instanceOf[EndorsementDateView]
+      val view = injector.instanceOf[EndorsementLocationView]
 
       status(result) mustEqual OK
 
       contentAsString(result) mustEqual
-        view(filledForm, mrn, index, mode)(request, messages).toString
+        view(filledForm, mrn, country.description, mode, index)(request, messages).toString
     }
 
     "must redirect to the next page when valid data is submitted" in {
 
-      setExistingUserAnswers(emptyUserAnswers)
+      val userAnswers = emptyUserAnswers.setValue(endorsement.EndorsementCountryPage(index), country)
+
+      setExistingUserAnswers(userAnswers)
 
       when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
 
-      val request = FakeRequest(POST, endorsementDateRoute)
-        .withFormUrlEncodedBody(
-          "value.day"   -> date.getDayOfMonth.toString,
-          "value.month" -> date.getMonthValue.toString,
-          "value.year"  -> date.getYear.toString
-        )
+      val request = FakeRequest(POST, endorsementLocationRoute)
+        .withFormUrlEncodedBody(("value", "test string"))
 
       val result = route(app, request).value
 
@@ -114,28 +109,30 @@ class EndorsementDateControllerSpec extends SpecBase with AppWithDefaultMockFixt
 
     "must return a Bad Request and errors when invalid data is submitted" in {
 
-      setExistingUserAnswers(emptyUserAnswers)
+      val userAnswers = emptyUserAnswers.setValue(endorsement.EndorsementCountryPage(index), country)
+
+      setExistingUserAnswers(userAnswers)
 
       val invalidAnswer = ""
 
-      val request    = FakeRequest(POST, endorsementDateRoute).withFormUrlEncodedBody(("value", ""))
+      val request    = FakeRequest(POST, endorsementLocationRoute).withFormUrlEncodedBody(("value", ""))
       val filledForm = form.bind(Map("value" -> invalidAnswer))
 
       val result = route(app, request).value
 
       status(result) mustEqual BAD_REQUEST
 
-      val view = injector.instanceOf[EndorsementDateView]
+      val view = injector.instanceOf[EndorsementLocationView]
 
       contentAsString(result) mustEqual
-        view(filledForm, mrn, index, mode)(request, messages).toString
+        view(filledForm, mrn, country.description, mode, index)(request, messages).toString
     }
 
     "must redirect to Session Expired for a GET if no existing data is found" in {
 
       setNoExistingUserAnswers()
 
-      val request = FakeRequest(GET, endorsementDateRoute)
+      val request = FakeRequest(GET, endorsementLocationRoute)
 
       val result = route(app, request).value
 
@@ -148,13 +145,8 @@ class EndorsementDateControllerSpec extends SpecBase with AppWithDefaultMockFixt
 
       setNoExistingUserAnswers()
 
-      val request =
-        FakeRequest(POST, endorsementDateRoute)
-          .withFormUrlEncodedBody(
-            "value.day"   -> date.getDayOfMonth.toString,
-            "value.month" -> date.getMonthValue.toString,
-            "value.year"  -> date.getYear.toString
-          )
+      val request = FakeRequest(POST, endorsementLocationRoute)
+        .withFormUrlEncodedBody(("value", "test string"))
 
       val result = route(app, request).value
