+++ test/pages/incident/endorsement/EndorsementDatePageSpec.scala
@@ -14,10 +14,8 @@
  * limitations under the License.
  */
 
-package pages.incident
+package pages.incident.endorsement
 
-import models.Index
-import org.scalacheck.Arbitrary
 import pages.behaviours.PageBehaviours
 
 import java.time.LocalDate
@@ -25,12 +23,6 @@ import java.time.LocalDate
 class EndorsementDatePageSpec extends PageBehaviours {
 
   "EndorsementDatePage" - {
-
-    implicit lazy val arbitraryLocalDate: Arbitrary[LocalDate] = Arbitrary {
-      datesBetween(LocalDate.of(1900, 1, 1), LocalDate.of(2100, 1, 1))
-    }
-
-    val index: Index = Index(0)
 
     beRetrievable[LocalDate](EndorsementDatePage(index))
